Moves story files
